# Fix #1: Setup Docker environment with Python/FastAPI