--- a/.env.example
+++ b/.env.example
@@ -1,0 +1,24 @@
+# Dialtone Configuration Template
+# Copy this file to .env and update values
+
+# Application settings
+APP_NAME=Dialtone
+APP_VERSION=0.1.0
+LOG_LEVEL=INFO
+
+# Paths
+OBSIDIAN_VAULT_PATH=/path/to/your/obsidian/vault
+
+# Processing limits
+MAX_UPLOAD_SIZE=52428800  # 50MB in bytes
+PROCESSING_TIMEOUT=35     # seconds
+MAX_CONCURRENT_REQUESTS=3
+
+# Server settings
+API_HOST=0.0.0.0
+API_PORT=8000
+WORKERS=1
+
+# Development settings
+RELOAD=false
+DEBUG=false