--- a/.gitignore
+++ b/.gitignore
@@ -1,0 +1,43 @@
+# Python
+__pycache__/
+*.py[cod]
+*$py.class
+*.so
+.Python
+venv/
+env/
+ENV/
+*.egg-info/
+dist/
+build/
+.pytest_cache/
+.mypy_cache/
+.coverage
+htmlcov/
+*.log
+
+# IDE
+.vscode/
+.idea/
+*.swp
+*.swo
+*~
+.DS_Store
+
+# Environment
+.env
+.env.local
+.env.*.local
+
+# Docker
+*.pid
+
+# Project specific
+logs/
+obsidian-vault/
+audio_uploads/
+temp/
+
+# Claude Dev Workflow - ignore working files, keep context
+workflow/logs/
+workflow/scratchpad/