--- a/app/__init__.py
+++ b/app/__init__.py
@@ -1,0 +1,3 @@
+"""Dialtone - Voice to Obsidian API."""
+
+__version__ = "0.1.0"