--- a/app/config.py
+++ b/app/config.py
@@ -1,0 +1,59 @@
+"""Configuration management for Dialtone."""
+
+import logging
+import sys
+from logging.handlers import RotatingFileHandler
+from pathlib import Path
+
+from pythonjsonlogger import jsonlogger
+
+from app.core.settings import settings
+
+
+def setup_logging() -> None:
+    """Configure application logging."""
+    log_level = getattr(logging, settings.log_level)
+    
+    # Create logs directory
+    log_dir = Path("logs")
+    log_dir.mkdir(exist_ok=True)
+    
+    # Configure JSON formatter
+    json_formatter = jsonlogger.JsonFormatter(
+        "%(timestamp)s %(level)s %(name)s %(message)s",
+        rename_fields={"timestamp": "@timestamp", "level": "severity"},
+    )
+    
+    # Console handler
+    console_handler = logging.StreamHandler(sys.stdout)
+    console_handler.setFormatter(json_formatter)
+    console_handler.setLevel(log_level)
+    
+    # File handler with rotation
+    file_handler = RotatingFileHandler(
+        log_dir / "dialtone.log",
+        maxBytes=10 * 1024 * 1024,  # 10MB
+        backupCount=5,
+    )
+    file_handler.setFormatter(json_formatter)
+    file_handler.setLevel(log_level)
+    
+    # Configure root logger
+    root_logger = logging.getLogger()
+    root_logger.setLevel(log_level)
+    root_logger.addHandler(console_handler)
+    root_logger.addHandler(file_handler)
+    
+    # Reduce noise from libraries
+    logging.getLogger("uvicorn.access").setLevel(logging.WARNING)
+    logging.getLogger("httpx").setLevel(logging.WARNING)
+    
+    logger = logging.getLogger(__name__)
+    logger.info(
+        "Logging configured",
+        extra={
+            "log_level": settings.log_level,
+            "app_name": settings.app_name,
+            "app_version": settings.app_version,
+        },
+    )