--- a/requirements-dev.txt
+++ b/requirements-dev.txt
@@ -1,0 +1,17 @@
+# Include production requirements
+-r requirements.txt
+
+# Development tools
+black==23.12.1
+mypy==1.8.0
+isort==5.13.2
+flake8==7.0.0
+
+# Testing
+pytest==7.4.4
+pytest-asyncio==0.23.3
+pytest-cov==4.1.0
+httpx==0.26.0  # For testing API
+
+# Type stubs
+types-aiofiles==23.2.0.0