--- a/requirements.txt
+++ b/requirements.txt
@@ -1,0 +1,17 @@
+# Core dependencies
+fastapi==0.109.0
+uvicorn[standard]==0.25.0
+pydantic==2.5.3
+pydantic-settings==2.1.0
+python-multipart==0.0.6
+
+# Async support
+httpx==0.26.0
+aiofiles==23.2.1
+
+# Logging and monitoring
+python-json-logger==2.0.7
+
+# Future AI dependencies (commented for now)
+# openai-whisper==20231117
+# ollama==0.1.7