# Development Guide: Setup Docker Environment with Python/FastAPI

## Overview
This guide provides step-by-step implementation for issue #1, establishing the Docker environment and FastAPI foundation for the Dialtone voice-to-Obsidian system.

## Pre-Implementation Checklist
- [ ] Ensure Docker and Docker Compose are installed
- [ ] Verify Python 3.11+ is available
- [ ] Check that git is initialized in the repository
- [ ] Confirm access to create files in the project directory

## Implementation Steps

### Step 1: Create Project Structure
Create the following directory structure:
```
dialtone/
├── docker-compose.yml
├── Dockerfile
├── requirements.txt
├── requirements-dev.txt
├── .env.example
├── app/
│   ├── __init__.py
│   ├── main.py
│   ├── config.py
│   ├── api/
│   │   ├── __init__.py
│   │   └── health.py
│   └── core/
│       ├── __init__.py
│       └── settings.py
├── tests/
│   ├── __init__.py
│   └── test_health.py
└── scripts/
    └── setup.sh
```

### Step 2: Docker Configuration

#### 2.1 Create Dockerfile
Multi-stage build for optimized production image with security best practices.

#### 2.2 Create docker-compose.yml
Define the voice-notes-api service with proper resource limits and health checks.

#### 2.3 Create .env.example
Template for environment variables following 12-factor app principles.

### Step 3: FastAPI Application

#### 3.1 Core Application Setup (app/main.py)
- FastAPI app factory pattern
- Middleware configuration
- CORS setup for PWA frontend
- API versioning

#### 3.2 Configuration Management (app/config.py, app/core/settings.py)
- Environment-based configuration
- Pydantic settings for type safety
- Validation of required settings

#### 3.3 Health Check Endpoint (app/api/health.py)
- Basic health status endpoint
- Version information
- System readiness checks

### Step 4: Development Tools

#### 4.1 Dependencies (requirements.txt, requirements-dev.txt)
- Production dependencies with pinned versions
- Development tools (black, mypy, pytest)

#### 4.2 Testing Setup (tests/test_health.py)
- Basic pytest configuration
- Health endpoint test
- Test coverage setup

#### 4.3 Setup Script (scripts/setup.sh)
- One-click setup for developers
- Dependency installation
- Initial configuration

### Step 5: Quality Assurance

#### 5.1 Code Formatting
- Black configuration (pyproject.toml)
- Import sorting with isort

#### 5.2 Type Checking
- mypy configuration
- Type hints on all functions

#### 5.3 Git Configuration
- .gitignore for Python/Docker artifacts
- Pre-commit hooks setup (optional)

## Code Implementation

### File-by-File Implementation Guide

1. **Dockerfile**: Multi-stage build with security hardening
2. **docker-compose.yml**: Service orchestration with resource limits
3. **requirements.txt**: Core dependencies (FastAPI, uvicorn, pydantic)
4. **app/main.py**: FastAPI application factory
5. **app/config.py**: Configuration management
6. **app/core/settings.py**: Pydantic settings model
7. **app/api/health.py**: Health check endpoint
8. **tests/test_health.py**: Basic health check test
9. **.env.example**: Environment variable template
10. **scripts/setup.sh**: Developer setup automation

## Testing Strategy

### Unit Tests
- Test health endpoint returns 200 OK
- Test configuration loading
- Test API documentation generation

### Integration Tests
- Docker Compose stack starts successfully
- All services healthy within 30 seconds
- Resource limits enforced

### Manual Testing
1. Run `docker-compose up --build`
2. Access http://localhost:8000/health
3. Check API docs at http://localhost:8000/docs
4. Verify logs show proper startup

## Documentation Updates

### README.md additions:
- Quick start section
- Docker setup instructions
- Development workflow
- Troubleshooting guide

### API Documentation:
- OpenAPI specs auto-generated
- Health endpoint documented
- Version information included

## Pull Request Checklist

Before creating the PR:
- [ ] All files created according to structure
- [ ] Docker build succeeds
- [ ] Tests pass with >80% coverage
- [ ] Black/mypy checks pass
- [ ] Health endpoint responds <50ms
- [ ] Memory usage <500MB
- [ ] Setup script tested on fresh system
- [ ] Documentation updated

## Commit Strategy

Suggested commit messages (conventional commits):
1. `feat: add Docker configuration for voice-notes-api`
2. `feat: implement FastAPI application structure`
3. `feat: add health check endpoint with tests`
4. `feat: add development setup scripts`
5. `docs: add quickstart and setup documentation`

## Performance Benchmarks

After implementation, establish baselines:
- Container startup time: Target <5s
- Health check response: Target <50ms
- Memory at idle: Target <500MB
- CPU at idle: Target <5%

## Security Considerations

Implemented security measures:
- Non-root user in container
- Resource limits to prevent DoS
- Input validation on all endpoints
- No hardcoded secrets
- Least privilege principle

## Next Steps

After PR is merged:
1. Run performance baseline tests
2. Document actual metrics
3. Prepare for issue #2 (audio upload endpoint)
4. Set up CI/CD pipeline (issue #6)

---
*Generated by Software Engineer Agent for Dialtone project*